gan-model/requirements.txt
# GAN Model - Image Generation Requirements

# Core Deep Learning Framework
tensorflow>=2.13.0
keras>=2.13.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Numerical Computing
numpy>=1.24.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Utilities
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Optional: GPU Support
# tensorflow-gpu>=2.13.0  # Uncomment if using GPU
